workflows: Fix libclc build failure on MacOS after 4bd3f3759259548e159aeba5c76efb9a0864e6fa
This is just a workaround.  I think the problem is that libclc is pulling
the lzstd library name from llvm-config, but isn't also getting its
install path (/usr/local/lib/).  I think the correct way to fix this
is for libclc to start using the cmake files instead of llvm-config.